+++ src/main/java/DAO/UserDAO.java
@@ -1,0 +1,47 @@
+package DAO;
+
+import DB.DBconnection;
+import DTO.Types;
+import DTO.userdto;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class UserDAO {
+
+    public int registerUser(userdto user) throws SQLException {
+        Connection connection = null;
+        PreparedStatement statement = null;
+        ResultSet resultSet = null;
+        int userId = 0;
+
+        try {
+            connection = DBconnection.getConnection();
+            String userSql = "INSERT INTO Users (name, email, password, userType) VALUES (?, ?, ?, ?)";
+            statement = connection.prepareStatement(userSql, PreparedStatement.RETURN_GENERATED_KEYS);
+            statement.setString(1, user.getName());
+            statement.setString(2, user.getEmail());
+            statement.setString(3, user.getPassword());
+            statement.setString(4, user.getUserType().name());
+
+            int affectedRows = statement.executeUpdate();
+
+            if (affectedRows > 0) {
+                resultSet = statement.getGeneratedKeys();
+                if (resultSet.next()) {
+                    userId = resultSet.getInt(1);
+                }
+            }
+        } finally {
+            try {
+                if (resultSet != null) resultSet.close();
+                if (statement != null) statement.close();
+                if (connection != null) connection.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }
+        return userId;
+    }
+}
