+++ src/main/java/DB/DBconnection.java
@@ -1,21 +1,29 @@
 package DB;
 
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.Properties;
 
 public class DBconnection {
-    private static final String JDBC_URL = "jdbc:mysql://localhost:3306/food";
-    private static final String USERNAME = "root";
-    private static final String PASSWORD = "root";
+    private static final Properties properties = new Properties();
+
+    static {
+        try (FileInputStream input = new FileInputStream("/database.properties")) {
+            properties.load(input);
+            Class.forName(properties.getProperty("jdbc.driverClassName"));
+        } catch (IOException | ClassNotFoundException e) {
+            throw new ExceptionInInitializerError("Failed to load database properties or driver");
+        }
+    }
 
     public static Connection getConnection() throws SQLException {
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-            return DriverManager.getConnection(JDBC_URL, USERNAME, PASSWORD);
-        } catch (ClassNotFoundException e) {
-            throw new SQLException("Failed to load MySQL JDBC driver", e);
-        }
+        String url = properties.getProperty("jdbc.url");
+        String username = properties.getProperty("jdbc.username");
+        String password = properties.getProperty("jdbc.password");
+        return DriverManager.getConnection(url, username, password);
     }
 
     public static void closeConnection(Connection connection) {
@@ -23,7 +31,7 @@ public class DBconnection {
             try {
                 connection.close();
             } catch (SQLException e) {
-                System.err.println("Failed connection " + e.getMessage());
+                System.err.println("Failed to close connection: " + e.getMessage());
             }
         }
     }
