+++ src/main/java/DB/database.properties
@@ -1,0 +1,4 @@
+jdbc.url=jdbc:mysql://localhost:3306/food
+jdbc.username=root
+jdbc.password=root
+jdbc.driverClassName=com.mysql.cj.jdbc.Driver
