+++ src/main/java/DTO/Types.java
@@ -1,0 +1,7 @@
+package DTO;
+
+public enum Types {
+	Retailer,
+	Consumer,
+	Charity
+}
