--- src/main/java/DTO/types.java
@@ -1,7 +1,0 @@
-package DTO;
-
-public enum Types {
-	Retailer,
-	Consumer,
-	Charity
-}
