+++ src/main/webapp/login.jsp
@@ -88,11 +88,9 @@
     <h2>LOGIN</h2>
     
     <form action="LoginServlet" method="post">
-        <label for="email">Email:</label>
-        <input type="email" id="email" name="email" required><br>
+        <input type="email" id="email" name="email" placeholder="Email" required><br>
         
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required><br>
+        <input type="password" id="password" name="password" placeholder="Password" required><br>
         
         <button type="submit">Login</button>
     </form>
